Add support for running tagged playbooks
The playbooks endpoint had been extended with a
tags 'node', so playbooks can be run in part.

Running a tagged playbook just uses env/cmdline
override with --tags {tag names} syntax. Since
this requires the file to be updated, code uses
a mutex to control access to the cmdline file in
the event multiple playbooks are competing for the
file.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>