Initial usage of the runner_cache
Now when a playbook starts it gets an entry in the
runner_cache, which can then be checked when the
caller requests a cancel. The entry points to the
runner object, and has a placeholder for the
"current_task". Current task will be added once
ansible_runners event handler provides the ident
of the event.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>