Add option to set explicit extravars
This PR add option to startplaybook endpoint which explicitly create
empty extravars file, which will enforce the ansible-playbook command to
use extravars as command line parameters, and not to be used from
extravars file.

If two parallel proccess are executed with different extravars specified
and single extravars file is used, it may happen that both proccess uses
same file, with same extravars, which is wrong.